Fix SDL warnings in CPU EP (#9975)

+++ onnxruntime/core/providers/cpu/rnn/rnn_helpers.h
@@ -110,7 +110,7 @@ void ReverseSequence(gsl::span<const T> inputs,
   for (int i = 0; i < batch_size; i++) {
     int seq_len = sequence_lengths[i];
 
-    for (int j = 0; j < seq_len; j++) {
+    for (ptrdiff_t j = 0; j < seq_len; j++) {
       gsl::span<const T> src = inputs.subspan(j * batch_size * input_size + i * input_size, input_size);
       gsl::span<T> dest = inputs_reverse.subspan(num_directions * (seq_len - j - 1) * batch_size * input_size + i * input_size, input_size);
